feat: Donation, balance & refund (#23)

+++ packages/Client/sdk/services/CampaignService.ts
@@ -1,6 +1,6 @@
-import { API, Campaign } from 'sdk'
+import { API, Service, Campaign, PaymentMethodService, CampaignPaymentMethod, PaymentMethodIdentifier } from 'sdk'
 
-export class CampaignService {
+export class CampaignService extends Service {
 	static get(id: number) {
 		return API.get<Campaign>(`/campaign/${id}`)
 	}
@@ -9,13 +9,41 @@ export class CampaignService {
 		return API.get<Array<Campaign>>('/campaign')
 	}
 
-	static save(campaign: Campaign) {
+	static getBalance(campaignId: number) {
+		return API.get<number>(`/campaign/${campaignId}/balance`)
+	}
+
+	static getShare(campaignId: number) {
+		return API.get<number>(`/campaign/${campaignId}/share`)
+	}
+
+	static async getDonationData(campaignId: number) {
+		const response = await API.get<Record<PaymentMethodIdentifier, string>>(`/campaign/${campaignId}/donation-data`)
+		return new Map(Object.entries(response)) as Map<PaymentMethodIdentifier, string>
+	}
+
+	static async declareAllocationPhase(campaignId: number) {
+		await API.post(`/campaign/${campaignId}/declare-allocation-phase`)
+	}
+
+	static async allocate(campaignId: number, destinationByPaymentMethodIdentifier: Map<string, string>) {
+		await API.post(`/campaign/${campaignId}/allocate`, destinationByPaymentMethodIdentifier)
+	}
+
+	static async save(campaign: Campaign, includeAllDonationChannels = false) {
+		if (includeAllDonationChannels) {
+			const paymentMethodIdentifiers = await PaymentMethodService.getAllIdentifiers()
+			campaign.activatedPaymentMethods = paymentMethodIdentifiers.map(identifier => new CampaignPaymentMethod(identifier))
+		}
 		return campaign.id
 			? API.put<Campaign>(`/campaign/${campaign.id}`, campaign)
 			: API.post<Campaign>('/campaign', campaign)
 	}
 
 	static delete(id: number) {
-		return API.delete(`/campaign/${id}`)
+		return this.confirmDeletion(() => API.delete(`/campaign/${id}`), {
+			heading: 'Delete Campaign',
+			content: 'Are you sure you want to delete this campaign irreversibly? All donations will be refunded.',
+		})
 	}
 }
